Add onDemand direct support and dispose operation for read-model facade

+++ packages/readmodel-adapters/resolve-readmodel-memory/src/reset.js
@@ -1,14 +1,23 @@
-import { hash } from './utils';
+import hash from './hash';
 
 export default function reset(repository, onDemandOptions) {
-    const key = hash(onDemandOptions);
+    if (!onDemandOptions) {
+        repository.forEach(({ collectionMap }) =>
+            collectionMap.forEach((collection) => {
+                collection.resetIndexes();
+                collection.remove({});
+            })
+        );
+        repository.clear();
+        return;
+    }
 
+    const key = hash(onDemandOptions);
     if (repository.has(key)) {
         repository.get(key).collectionMap.forEach((collection) => {
             collection.resetIndexes();
             collection.remove({});
         });
-        repository.get(key).onDestroy();
         repository.delete(key);
     }
 }
